refactor: extract inference_parallel.py and job_helpers.py
Split backend/service/inference.py (634 -> 384 lines):
- inference_parallel.py: ParallelInferenceMixin with threaded pipeline

Extract from ui/workers/gpu_job_worker.py (581 -> 535 lines):
- job_helpers.py: create_job_snapshot() standalone function

split_view.py (721) left as-is — too tightly coupled for clean split.
All 427 tests pass.
